Basic MSG to JSON functionality
When a .msg is loaded into the file editor, the file editor requests a JSON representation from MSGFormat and writes it to a file.

Alongside this, I've cleaned up some names and privacy modifiers so that the created JSON file looks how I'd like it to.

As an implementation change, function arguments are now shorts rather than ushorts to accommodate the obvious -1s that are present in many files. This shouldn't affect any current functionality; the only thing using Args is InitializeText, and the color arguments are from 0 - 255.