Load an already-processed file from JSON
If a working file (current.json) exists in the MSG file's working directory, it is loaded instead of the version from the PAR file.

Because I made messages' function arrays internal rather than public, I have to pull that information out of the human-readable string version. It cuts down vertical bloat in the JSON file, I do want to keep it.

Note to self: line 251 is crucial for reassignment back to the array, otherwise the array will remain null. Writing Sections[s].Headers[h].Messages[m] throughout is wordy, but don't be a dummy.