Proper edit counting; write edited message strings to JSON
Edits are now properly tracked and accounted for via a list of the controls that have been edited. Controls add or remove themselves from that list at need.

Edited import textboxes now properly record their edits back into the fields of MSGFormat, which then write properly to JSON, and load properly from JSON.


There is a bug involved with this new change. InitializeText currently uses the values in the functions of the message to append the appropriate text from the message string (ignoring Color tags) while simultaneously colouring the text to match the position of the tags. But the functions in JSON only reflect the functions for the import string, not the export string.

One way of resolving this would be to keep and save a second set of functions for the export string. This could be done by loading it from orig.json (which could then be extrapolated to also always loading the export string itself from orig.json, and doing away with the distinct export and import strings in every json file). Or two sets of functions could be saved in every json file, which I definitely don't want to do.

Another resolution, which might be cleaner, would be to colour text in InitializeText based on the positions of Color tags rather than functions. Only the import needs to be reflected in the functions, since the functions are what's going back into the MSG file for the game to read. And only the color needs to be reflected in the export, no other functionality.