orig.json is no longer created; EditedControls bugfix
When saving changes to JSON for the first time, ver0001.json (the first current.json) would always be identical to orig.json. That file is no longer created, and version numbers start at 0 to better indicate the first backup's purpose.

-----

Because textbox edits are set up to utilize the LostFocus event, the currently focused textbox on close wouldn't count as an edit until close. This would also cause the textbox to be blank on reload.

The fix for the above bug was to give the file editor form focus at the start of the FormClosing event, which will cause the LostFocus event to occur, updating EditedControls if necessary for the remainder of the FormClosing event to make use of.