Bugfix: "Can't expand stream" on files larger than original
In Convert, using new MSGFormat (source.Stream, 0, source.Stream.Length) filters down to DataStream (stream, offset, length), which sets canExpand to false.

In order to fix this, I first tried using new MSGFormat (source.Stream), and then I realized a bit of an issue. The stream from the ParFile source is actually being disposed of, even though it's being passed to DataStream's constructors and referenced by fields within DataStream.

The difference is that the three-argument constructor sets the passed stream as ParentDataStream, and BaseStream as stream.BaseStream. ParentDataStream is the stream that gets disposed of, so this never became an issue. The one-argument constructor sets the passed stream as BaseStream, so the disposal becomes immediately obvious.

The disposal occurs in Node's TransformWith's call to CastAndChangeFormat, which calls ChangeFormat, which assumes you want disposal of the stream. And that makes sense, it's just at odds with the single-arg constructor of DataStream.

At any rate, I should consider keeping the fixed-length stream, in which case I would make the new stream in WriteToBin on demand, bypassing this issue. But this is fine.