Added ParFile fields "WasCompressed", "CompressionVersion"
It will be useful for my purposes to know if a file was compressed in the loaded PAR, and also with what version it was compressed. I want to be able to compress files in the PAR, or to not do so, based on whether it was compressed to begin with.