More meaningful file edit counting
Instead of checking against their CanUndo field, the original textbox text is now stored as the Tag of the FlowLayoutPanel that contains the textbox. On TextChanged, the textbox Text is checked against that Tag. This way, if the text is the same as it was when the file was opened, but CanUndo is true (e.g. delete and re-type those same deleted characters) it will still be counted as being unedited.

UpdateSpeaker is still performed on LostFocus, since it would be unnecessary to do it more often.

UpdateTitle() in FileEditorForm adds or removes the title's asterisk based on the contents of EditedControls. It is called to end the TextChanged delegates.