Added automatic per-file compression in ParArchiveWriter
If ParArchiveWriterParameters is passed with CompressorVersion equal to 3 (or greater, at present), then ParArchiveWriter's CompressFiles function will perform compression using each file's compression version from when it was initially decompressed, i.e. on initial load from the PAR.

There's an additional catch-case in Compressor, which will utilize the source ParFile's CompressionVersion if the CompressionParameters passed to it have a Version > 2.


The field being used, ParFile.CompressionVersion, should currently be initialized to 0 everywhere, except at time of decompression where it will be set to the version used in the decompression. If the file was never compressed to begin with, the ParFile field WasCompressed should be set to false. It would be an error for WasCompressed to be true and CompressionVersion to be 0.