First pass of "Save PAR As..." implementation
In this first pass, I'm only testing loading a PAR and then immediately saving that same PAR.

The generated PAR has two problems in its 16-byte header. File size is being written as 0, and another byte is different whose purpose I don't immediately recall.

The newly-written file isn't being properly let go of until program close / the Dispose method being called on the Node. That's probably because ParArchiveWriter assigns the open writing stream as the new ParFile's stream. I'll have to edit ParArchiveWriter to handle this and the header problems.


Also, I want the program to keep its state after a save. But because TransformWith is destructive, I need to do one of two things to accomplish that. Either:

- (1) Continue to use ParArchiveWriter to transform, but clone the Node before converting that clone via TransformWith. YARHL has the ability to clone at present, but the most recent release was in 2020 when that wasn't implemented yet, so to do this I'd have to switch from NuGet to building YARHL myself. That would also mean coming to terms with whatever changes YARHL has had since 2020.

- (2) Separate the workings of Convert out into a second method that Convert calls (void Write?). Then, call that second method myself to write the file to disk, rather than using TransformWith.


It's hard for me to imagine the situation in which I would actually want to convert away from NodeContainerFormat, since the state of this program is meant to persist. Getting clones only to use Convert on them feels very unnecessary for my purposes, which are to write the file to disk and then seamlessly continue operation. So I'd probably like to do option 2, get everything up and running, and then update YARHL anyway after everything else is functioning correctly.

As a note to myself, it would be nice to have a NodeContainerFormat -> NodeContainerFormat converter that simply compresses the files in the container. That seems to fall more in line with the ParFile -> ParFile Convert in Compressor and Decompressor.