ParArchiveWriter now keeps tags from ParArchiveReader
Somewhere in the shuffle through TransformWith, the root Node created in ParArchiveReader gets replaced, and the Tags generated go with it. I chose to pass a Dictionary in ParArchiveReaderParameters, which gets filled in the same manner, but this way the Tags are kept.

Another way of doing this would have been to attach these Tags to the top-level folder Node in the archive (folder[0] in ParArchiveReader's Convert), but that's hacky.