Bugfix: Proper disposal of stream after writing
Added a using directive to make sure the stream was disposed of when leaving scope. Manually calling Dispose at the end of Convert also works just fine.

The second part of the fix is writing the stream to a second stream intended to be output by Convert, and keeping the data alive via that new stream.

It is worth noting that this error only occurred when OutputPath was set in the parameters, since it was only in that circumstance that the stream was written to a file.