ParArchiveWriter.Convert split into two functions
As discussed in the initial "Save PAR As..." implementation commit, I've split the workhorse part of Convert off into its own function, WriteArchive. Convert calls WriteArchive, which returns the DataStream for Convert to wrap as a new ParFile.

I've swapped the order of operations on the two streams, as well. Everything is written to memory, then if OutputPath is set, a second stream is created that writes out to disk. The first stream is the one returned.

WriteArchive is also static, so that I can call it directly in MainForm to write the new archive out to file.