Just-right files allowed to fit into PAR block
In ParArchiveWriter.WriteFiles, a new block was created when the length of the next file to be inserted was equal to the remaining length of the block. In other words, a file whose added length would bring the block in alignment with 2048 was pushed to the next block. I've changed this so that these files fit into the block, filling it up.

I can only speak for Kenzan, but I'm pretty certain this works just fine as a space-saving technique. Of course, if this needs to be reverted down the line, it's only a one-character edit.