Remove file writing as a side effect in ParArchiveWriter
It's somewhat convenient that ParArchiveWriter writes the file out to disk when it's done converting the Node, but it doesn't fall in line with YARHL's workflow, and it hides the file write away from MainForm.

What ParArchiveWriter now does is simply convert a NodeContainerFormat that's full of ParFiles to a single ParFile acting as an archive.

I then write the stream out to disk in MainForm, but it could conceivably be used for other purposes. For example, ParArchiveWriter could now be used recursively in a way it couldn't have been before, because of that side effect of writing the finished PAR to disk.