Basic file inclusion in PAR functionality
IncludeFile swaps out the old Format of a Node for the new Format, preparing the state of the PAR for when it is saved. This will save the new stream in the saved PAR instead of the old stream. The old stream is retained via a list of replaced Formats. The new stream is also written out to disk for storage.

When the program closes, the current list of included paths is written out to a text file. If this PAR is reopened by the program, those inclusions will be loaded from the files that were written by IncludeFile.

What remains to do is to write the inverse of IncludeFile, as well as allowing the user access to this functionality rather than performing inclusions automatically.