Files loaded in the File Editor will not be reloaded on form's open
The File Editor was performing a Load on every opening of the form. That's unnecessary, the RGGFormat will stay in memory and can be reused so long as it's attached to the Node.

Having this work correctly means squashing a bug related to the speaker textbox's LostFocus event being called twice, the second time being with an empty string. What I've done is disallow empty strings, since I don't want to propagate an empty string through the file anyway. (I've also moved the check into the LostFocus delegate, and I'm not updating Misc until ProcessEdits, the same as for message text.)

It also requires the controls to be generated anew on every open of the File Editor.