Version selector uses RGGFormat instead of JSON DataStreams
The version selector now swaps around formats instead of streams. This should save a little work, because the format is already in place with all its fields. There is still a bit of a pause on larger files when swapping between versions, which must be because of UpdateControls.

The controls currently persist between File Editor openings of the same file, even when not saving. I think I like this behaviour, for the most part.

It would probably mitigate the length of that pause if I give each Version its own set of persistent generated controls, and swap between them on the fly, since that way I won't have to interpret Color tags. Or another way to tackle it within UpdateControls might be to fill in all the text visually, and then do a second pass for just the InitializeText calls.

If the former, I wonder if it's better to have them all in Controls but set Visible to false, or if I should add/remove them from Controls as needed.