Delay CopyFormat calls until absolutely necessary
The current state of the file will persist in the current Format of the Node. It doesn't need to be copied immediately upon load, whether from the PAR or from disk.

CopyFormat is now only called to retain the current Format. If the version of the file is changed via the selector, the current format of the file is stored in the LoadedVersions array.