Inclusions rework
Inclusions are now more closely integrated with the Formats generated by RGGFormat and the File Editor. Rather than storing the included Stream and the replaced Format, ProjectManager just keeps a list of included Nodes. Each Format gets its own Stream, so when it comes time to include that Format, it's only a matter of updating that Stream before doing so.

From a practical standpoint, including Nodes means changing Format to the included version's Format, and excluding a Node after inclusion means changing Format to the original Format. The attached Streams in the Formats ensure that the file, if saved, will come out as intended.