Fix: Inclusions not being initialized for/by the File Editor
Included versions will now work after boot when opened with the File Editor.

Note: There is a bug present in this version. Edits aren't recognized in any version of any file other than the original. This bug has actually been present throughout, because the controls' delegates are attuned to the original instance of the Format. When the copy gets a reference to Handle in CopyFormat, they're still acting and reacting based on the fields of the original instance, so edits work after a LoadFromJSON call, but don't work on swapped-to copies.

In other words, it seems that every Format copy needs its own controls. Luckily, I wanted to do that anyway.