Complete version-saving functionality
The Save Version button is a split button. The default option is to save a new version with a user-provided name. The other options are to save as a new version with a default name, or to overwrite the current version (which for now asks no questions).

I'll note that I think saving with a default name might make more sense as the default option, since it's quick (and will become painless when version renaming is implemented).