RGGFormat: EnableControls method and MSGFormat implementation
RGGFormat now has an EnableControls method. This method "enables" or "disables" the controls of the format so that they are no longer editable by the user. The colour passed in is the new back colour for the controls.

MSGFormat's implementation of EnableControls makes all import textboxes read-only or not as is necessary.

Importantly, MSGFormat sets its RGGFormat.Enabled field to true when created via Convert, and passes its Enabled value through to copies in CopyFormat. GenerateControls now acts based on Enabled to generate read-only controls if need be.