Change to inclusion/exclusion semantics
On inclusion from initial load, the format from the PAR is preserved in the FIrstLoadBuffer slot. The new format is no longer a created ParFile, but is instead generated by a CopyFormat call on the old format. The fields are loaded in from the version's JSON file, and because fields are necessary for CopyFormat to function, they are first loaded into the old format. (The content of the old format's fields don't matter at this time, its value here is in preserving the PAR's stream.)

CopyFormat was changed to accept a new stream as an argument. When this argument is passed, its contents are utilized for the creation of the copy, or as the copy's stream if uniqueStream is true. If the argument is null or not passed, CopyFormat works as it did prior.

When the file editor is opened on a file with an inclusion from initial load, the editor no longer loads the version JSON for the inclusion, since that is covered by the project manager.

In addition to the above, as a small semantic change, the contents of the IncludedVersion tag of a node will be handled exclusively by the project manager. IncludeFile now takes a versionTag argument that specifies the name of the version being included. (In the future, when the option exists to include a specific version of the file from the main form, ownership of the file version list will likely move to the node itself, and will be more comfortably handled by the project manager.)